pyproject.toml
[tool.poetry]
name = "home-work27-django"
version = "0.1.0"
description = ""
authors = ["GeFoGen <gefogen>"]
readme = "README.md"
packages = [{include = "home_work27_django"}]

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.1.2"
djangorestframework = "^3.14.0"
psycopg2 = "^2.9.4"
pillow = "^9.2.0"
djangorestframework-simplejwt = "^5.2.2"
pytest-django = "^4.5.2"
pytest-factoryboy = "^2.5.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
